+++ src/xyang/parser/statements/__init__.py
@@ -16,6 +16,12 @@ from .type import TypeStatementParser
 from .uses import UsesStatementParser
 from .must import MustStatementParser
 from .when import WhenStatementParser
+from .choice import ChoiceStatementParser
+from .grouping import GroupingStatementParser
+from .container import ContainerStatementParser
+from .list import ListStatementParser
+from .leaf import LeafStatementParser
+from .leaf_list import LeafListStatementParser
 
 __all__ = [
     "AnydataStatementParser",
@@ -34,4 +40,10 @@ __all__ = [
     "UsesStatementParser",
     "MustStatementParser",
     "WhenStatementParser",
+    "ChoiceStatementParser",
+    "GroupingStatementParser",
+    "ContainerStatementParser",
+    "ListStatementParser",
+    "LeafStatementParser",
+    "LeafListStatementParser",
 ]